+++ sys/drv/atkbdc.c
@@ -174,9 +174,9 @@ static int atkbdc_attach(device_t *dev) {
 static driver_t atkbdc_driver = {
   .desc = "AT keyboard controller driver",
   .size = sizeof(atkbdc_state_t),
+  .pass = SECOND_PASS,
   .probe = atkbdc_probe,
   .attach = atkbdc_attach,
-  .pass = SECOND_PASS,
 };
 
 DEVCLASS_ENTRY(pci, atkbdc_driver);
